!feat: Download today's user stories by location
That change is step one to expend our story database.
Now we take the users from StoryStory that uploaded a story with a location tag that we follow today.
Then we download all of their stories using instagram_bot.
This way, we will be able to collect all the stories that weren't tagged at the location, which we might have otherwise missed.
The next step is to create location folders on user's Drive main stories folder for better structure.